First version - crashes if you select a different tab
First version. Vibe coded over one day. Started the tutorial for textual and went nuts with the design.
Going to revisit once LLM calls are working/endpoint is established in tldw. At that point, will resume working on this and making this the default UI until webapp version comes along.